pyproject.toml
[tool.poetry]
name = "hack-ai-resume-bot"
version = "0.1.0"
description = ""
authors = ["Siddhant Rao <siddhant.rao@spit.ac.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
uagents = "^0.9.2"
joblib = "^1.3.2"
pypdf2 = "^3.0.1"
numpy = "^1.26.2"
scikit-learn = "^1.3.2"
torch = "^2.1.2"
transformers = "^4.36.2"
google-generativeai = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
